AI-Market-Analyst/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.5
distro==1.9.0
exceptiongroup==1.2.1
frozendict==2.4.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.54
lxml==5.2.1
marshmallow==3.21.2
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.26.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
peewee==3.17.3
pip==23.3.1
pydantic_core==2.18.2
pydantic==2.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
wheel==0.43.0
yarl==1.9.4
yfinance==0.2.38
